requirements.txt
SQLAlchemy~=2.0.40
fastapi~=0.115.12
python-jose~=3.4.0
pydantic-settings~=2.9.1
python-dotenv~=1.1.0
passlib~=1.7.4
pydantic~=2.11.4
pip~=25.1.1
cryptography~=44.0.3
setuptools~=65.5.0
typing_extensions~=4.13.2
websockets==10.0